requirements.txt
asgiref==3.7.2
boto3==1.34.9
botocore==1.34.9
branca==0.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
django-environ==0.11.2
django-filter==23.5
django-storages==1.14.2
djangorestframework==3.14.0
gpxpy==1.6.1
idna==3.6
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
numpy==1.26.2
Pillow==10.1.0
psycopg2==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.18
